requirements.txt
Flask==2.3.3
google-cloud-speech==2.21.0
google-cloud-storage==2.10.0
openai==1.3.0
python-dotenv==1.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
pydub==0.25.1
ffmpeg-python==0.2.0
